Cleanup IPC api structure
