+++ src/ipc/ipcHandlers.spec.ts
@@ -4,13 +4,15 @@ import { range } from 'lodash';
 import { v4 } from 'uuid';
 import ipcCommands from './ipcCommands';
 import addIpcHandlers from './ipcHandlers';
-import generateContract from './utils/documentGeneration/contractGeneration/generateContract';
-import { ContractGenerationArgs } from './utils/documentGeneration/contractGeneration/generateContractMarkdown';
-import { GenerateRentIncreasePdfArgs } from './utils/documentGeneration/rentIncrease/GenerateRentIncreasePdfArgs';
-import * as rentIncreaseModule from './utils/documentGeneration/rentIncrease/generateRentIncreasePdf';
+import generateContract from './modules/documentGeneration/contractGeneration/generateContract';
+import { ContractGenerationArgs } from './modules/documentGeneration/contractGeneration/generateContractMarkdown';
+import { GenerateRentIncreasePdfArgs } from './modules/documentGeneration/rentIncrease/GenerateRentIncreasePdfArgs';
+import * as rentIncreaseModule from './modules/documentGeneration/rentIncrease/generateRentIncreasePdf';
+import * as exportObjectModule from './modules/persistence/exportObject/exportObject';
+import * as importObjectModule from './modules/persistence/importObject/importObject';
+import * as uploadDocumentModule from './modules/persistence/uploadDocument/uploadDocument';
 import * as persistenceModule from './utils/persistence';
 import { DocumentTarget } from './utils/persistence/documentTarget';
-import * as uploadDocumentModule from './utils/persistence/uploadDocument';
 import MonthYear from '_/extensions/date/month_year.extension';
 import InvoiceBuilder from '_/test/builders/invoice.builder';
 import LandlordBuilder from '_/test/builders/landlord.builder';
@@ -19,10 +21,13 @@ import ResidentBuilder from '_/test/builders/resident.builder';
 import ResidentInvoiceInformationBuilder from '_/test/builders/residentInvoiceInformation.builder';
 import { ipcMain, ipcRenderer } from '_/test/electronModuleMock';
 
-jest.mock('./utils/documentGeneration/contractGeneration/generateContract', () => ({
-  __esModule: true,
-  default: jest.fn(),
-}));
+jest.mock(
+  './modules/documentGeneration/contractGeneration/generateContract',
+  () => ({
+    __esModule: true,
+    default: jest.fn(),
+  }),
+);
 
 describe('addIpcHandlers', () => {
   let exportObjectSpy: jest.SpyInstance;
@@ -33,9 +38,9 @@ describe('addIpcHandlers', () => {
 
   beforeEach(() => {
     exportObjectSpy = jest
-      .spyOn(persistenceModule, 'exportObject')
+      .spyOn(exportObjectModule, 'exportObject')
       .mockReturnValue();
-    importObjectSpy = jest.spyOn(persistenceModule, 'importObject');
+    importObjectSpy = jest.spyOn(importObjectModule, 'importObject');
     openDirectoryDialogSpy = jest.spyOn(
       persistenceModule,
       'openDirectoryDialog',
