+++ src/ipc/modules/documentGeneration/contractGeneration/generateContractMarkdown.spec.ts
