+++ src/ipc/modules/documentGeneration/contractGeneration/generateContractMarkdown.ts
